requirements.txt
customtkinter
uuid
keyboard
Pillow
pyserial
requests
ping3
pysdl2
websocket
websocket-client